Update page.tsx
The check to find a person using params from the current url path included the comparison between two values. One was a string (params.person) and one was a number (contact.id) assuming the error caused was because of the strict comparison between the wrong data types. I changed params.person to id inside of the comparison